// lib/faust-types.ts
export interface MissionFields {
  missionTitle1: string;
  missionTitleItalic: string;
  missionGoal: string;
  missionCtaText: string;
  missionCards: MissionCard[];
}

export interface Mission {
  missionFields: MissionFields;
}

export interface MissionData {
  missions: {
    nodes: Mission[];
  };
}